Replace expansive regex with find (#976)
std::regex is atrociously slow. We could replace it with another lib but
in this case we don't really require the power of regex expressions and
just need a regular string equality match.

Performance wise :
For 66 criterions and 247320 variables
Using regex, variablesGroups::search takes 74s out of 90 (82%)
Using find it takes 0.5s out of 20 (2,5%)