Weights not correctly read for subproblems (#990)
In problem generation, a file `weights.txt` is generated to map problem
names to their corresponding weights. Initially, this mapping was
constructed by fetching all mps files in the directory after the antares
step and extracting the year number from the file name.

In memory mode, Antares does not write mps file anymore. The weights
file generation is now based on the `problems_and_data` structure in
memory that gathers the problems and their year number. Moreover,
depending on the solver name, the filename written in the `weights.txt`
is now correct (.svf or .mps) depending on the context.