[ANT-2202] Vectorize row and col names change (only available with Xpress) in solver to accelerate problem generation (#994)
In problem generation, when reading the problem in memory from the
Antares data structure, row and column names were added one by one to
the `Problem` object. This is very inefficient, and is a major
bottleneck in problem generation time (verified with vtune).

Solver Xpress provides a method to vectorize changes in row and column
names. This PR exploits this vectorization to accelerate problem
generation (when Xpress is used). As this feature is not provided by
other solvers, no change in performance can be expected, we simply
update the solver interfaces accordingly.

---------

Co-authored-by: Jason Maréchal <45510813+JasonMarechal25@users.noreply.github.com>