[ANT-3671] Fix numerical issues in Benders (#986)
Several numerical issues can occur in Benders
- 
- [x] Fixing candidate values in the subproblems to the value obtained
from the master can lead to subproblem infeasibility if the master
solution is numerically slightly below the lb, above the ub, or is a not
rounded integer value. Also master solution that are close to zero given
a tolerance should be set to zero in the subproblems to avoid having
large range of coefficients in the subproblems leading to numerical
issues.
- [x] Subproblem gradients that are close to zero (given a tolerance)
should appear as zero in the cut added to the master to avoid numerical
issues in the master due to large range of coefficients values

---------

Co-authored-by: Jason Marechal <jason.marechal+github@gmail.com>