Feature/cut aggregation (#1126)
This PR is related to a new functionality implemented on the Benders
algorithm (the MPI version).

Core new functionality : 

Overview : 

An investment problem for grid expansion study is built on a master
problem (mps file) and subproblems (mps files) and options.json, a json
file that provides parameters for the study.
The development is concerned with the AGGREGATION attribute. This
attribute handles the cut added to the master problem at each iteration.
For no AGGREGATION is a boolean and we are supposed to have the
following behaviour:
If AGGREGATION is false : 
In this cas we are supposed to add as much cuts as subproblems
		Rhs = s*(z-z_cut) + alpha_i 
	     -    If AGGREGATION is true :   
In this case we are supposed to add one cut at each iteration, summing
constraints of each subproblem
		-sum(rhs) > s*(z-z_cut) +sum(alpha_i ), for each subproblem 
A bug has been found while developing the new functionality : Actually,
when AGGREGATION is true, we add as many cuts as the number of
processes.




`void BendersMpi::master_build_cuts(std::vector<SubProblemDataMap>
gathered_subproblem_map)
{
    SetSubproblemCost(0);

    for (const auto& subproblem_data_map: gathered_subproblem_map)
    {
for (auto&& [sub_problem_name, subproblem_data]: subproblem_data_map)
        {
SetSubproblemCost(GetSubproblemCost() +
subproblem_data.subproblem_cost);
            BoundSimplexIterations(subproblem_data.simplex_iter);
        }
    }

    _logger->display_message("\tSolving subproblems...");

    _data.ub = 0;

    for (const auto& subproblem_data_map: gathered_subproblem_map)
    {
        BuildCutFull(subproblem_data_map);
    }


_logger->LogSubproblemsSolvingCumulativeCpuTime(_data.subproblems_cumulative_cputime);
    _logger->LogSubproblemsSolvingWalltime(_data.subproblems_walltime);
}`

As we can see in the code above, we add a cut at each gathered set of
subproblem for each proc they have been solved on.

The issue is that the problem is written in this case in a way to get
one cut which lead to convergent error (the algorithm converges but to a
wrong value).

What has been changed : 

The main idea of this development is to make it possible to allow
configuring the number of cuts that will be aggregated at each
iteration.
AGGREGATION will be an integer that give the number of cuts to be added
at each iteration.
AGGREGATION can be any integer but if it’s superior to the number of
subproblem, we will only consider the number of subproblems as added
cuts per iteration.


How cuts will be created : 
The idea is to partition the subproblems to n_cut partition and build a
constraint by summing the righ parameters for the chosen subproblems in
the cut.
In the code, this process is handled by the function
split_subproblem_data.
This function will render a vector where each slot is a Map with the
subproblems that will build the cut.
We also add the method split_subproblem_data_with_shuffle. Basically it
does the same thing but it randomly select the subproblem to build the
cut.
For now we only use the first function but we kept the second one for
further performance study.

So we refactored the code the following way : 

`void BendersMpi::master_build_cuts(std::vector<SubProblemDataMap>
gathered_subproblem_map)
{
    SetSubproblemCost(0);

    for (const auto& subproblem_data_map: gathered_subproblem_map)
    {
for (auto&& [sub_problem_name, subproblem_data]: subproblem_data_map)
        {
SetSubproblemCost(GetSubproblemCost() +
subproblem_data.subproblem_cost);
            BoundSimplexIterations(subproblem_data.simplex_iter);
        }
    }

    _logger->display_message("\tSolving subproblems...");

    _data.ub = 0;

    if (_world.rank() == rank_0) 
    {
        if (_data.nsubproblem < _options.AGGREGATION) 
        {   
std::string logging_str = "AGGREGATION : " +
std::to_string(_options.AGGREGATION) + " is larger than the number of
subproblems : " + std::to_string(_data.nsubproblem) +
"setting AGGREGATION to " + std::to_string(_data.nsubproblem);
            _logger->display_message(logging_str) ; 
            std::cout<<logging_str<<std::endl ; 
            _options.AGGREGATION = _data.nsubproblem ; 
        } 
std::vector<SubProblemDataMap> subproblem_per_cut =
split_subproblem_data( gathered_subproblem_map, _options.AGGREGATION) ;
std::cout<<"subproblem_per_cut size
"<<subproblem_per_cut.size()<<std::endl ;
        for (const auto& subproblem_data_map: subproblem_per_cut)
        {
            BuildCutFull(subproblem_data_map);
        }
    }


_logger->LogSubproblemsSolvingCumulativeCpuTime(_data.subproblems_cumulative_cputime);
    _logger->LogSubproblemsSolvingWalltime(_data.subproblems_walltime);
}
`
`void BendersBase::BuildCutFull(const SubProblemDataMap&
subproblem_data_map)
{
    check_status(subproblem_data_map);
    compute_cut_aggregate(subproblem_data_map);
}`

Test protocol 

So the new changes has been tested by an end to end test that consider
the following study : _data_test/investment_study._

We added ; benders_aggreg_cuts.feature to run the test scenartion and
make sure that we converge to the right overall cost
__

---------

Co-authored-by: BOUCHEHDA Hedi (Externe) <bouchehdahed@gm0winl993.bureau.si.interne>
Co-authored-by: Thomas Bittar <thomas.bittar@rte-france.com>